Extract `request_leap_if_not_redundant()`
